update keyword logic and move hooks folder
